+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/logic/components/flow/GroupNode.tsx
@@ -10,6 +10,8 @@ import { useElementTitle, ElementProperties } from "@lib/hooks/useElementTitle";
 import { useTranslation } from "@i18n/client";
 import { FormElementTypes } from "@lib/types";
 
+import { useTreeRef } from "@formBuilder/components/shared/right-panel/treeview/provider/TreeRefProvider";
+
 const OptionRuleSvg = ({ title }: { title?: string }) => {
   return (
     <svg xmlns="http://www.w3.org/2000/svg" width={34} height={34} fill="none">
@@ -47,12 +49,14 @@ export const GroupNode = (node: NodeProps) => {
   const { t } = useTranslation("form-builder");
 
   const { getTitle } = useElementTitle();
+  const { togglePanel } = useTreeRef();
 
   const handleClick = {
     onClick: () => {
       setId(node.id);
       // Reset selected element id
       setSelectedElementId(0);
+      togglePanel && togglePanel(true);
     },
   };
 
@@ -81,7 +85,6 @@ export const GroupNode = (node: NodeProps) => {
         id={node.id}
         className={cn(
           "space-y-2 rounded-md border-2 border-indigo-500 p-4 text-white",
-          "space-y-2 rounded-md border-1 border-indigo-500 p-4 text-white",
           groupIsSelected
             ? "bg-violet-200 shadow-logicSelected"
             : "bg-gray-soft shadow-logicDefault",
@@ -97,7 +100,7 @@ export const GroupNode = (node: NodeProps) => {
               "absolute right-[-20px] top-[-20px] cursor-pointer outline-offset-8 outline-slate-800 hover:scale-125 rounded-full"
             )}
           >
-            <QuestionRuleSvg title={t("groups.editSection", { name: node.data.label.name })} />
+            <QuestionRuleSvg title={t("groups.editPage", { name: node.data.label.name })} />
           </button>
         )}
         {!node.data.children.length && <div className="min-h-[50px] min-w-[200px]"></div>}
@@ -156,6 +159,7 @@ export const GroupNode = (node: NodeProps) => {
                 evt.stopPropagation();
                 setId(node.id);
                 setSelectedElementId(Number(child.index));
+                togglePanel && togglePanel(true);
               }}
               className={cn(
                 nodeClassName,