+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/clientSide.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { NextButton } from "@clientComponents/forms";
+import { NextButton } from "@clientComponents/forms/NextButton/NextButton";
 import { useRouter } from "next/navigation";
 import { useTranslation } from "@i18n/client";
 import { FormRecord, TypeOmit } from "@lib/types";
@@ -33,6 +33,7 @@ export const FormWrapper = ({
       renderSubmit={({ validateForm, fallBack }) => {
         return (
           <NextButton
+            formRecord={formRecord}
             language={language as Language}
             validateForm={validateForm}
             fallBack={fallBack}