+++ lib/users.ts
@@ -126,7 +126,12 @@ export const getOrCreateUser = async ({
  */
 export const getUser = async (ability: UserAbility, id: string): Promise<AppUser> => {
   try {
-    checkPrivileges(ability, [{ action: "view", subject: "User" }]);
+    checkPrivileges(ability, [
+      {
+        action: "view",
+        subject: "User",
+      },
+    ]);
 
     const user = await prisma.user.findFirstOrThrow({
       where: {
@@ -174,11 +179,7 @@ export const getUsers = async (
     checkPrivileges(ability, [
       {
         action: "view",
-        subject: {
-          type: "User",
-          // Empty object to force the ability to check for any user
-          object: {},
-        },
+        subject: "User",
       },
     ]);
 
@@ -222,7 +223,12 @@ export const getUsers = async (
  */
 export const updateActiveStatus = async (ability: UserAbility, userID: string, active: boolean) => {
   try {
-    checkPrivileges(ability, [{ action: "update", subject: "User" }]);
+    checkPrivileges(ability, [
+      {
+        action: "update",
+        subject: "User",
+      },
+    ]);
 
     const [user, privilegedUser] = await Promise.all([
       prisma.user.update({